+++ R/class_future.R
@@ -74,6 +74,9 @@ future_class <- R6::R6Class(
       self$ensure_globals()
       globals <- self$globals
       globals$.targets_target_5048826d <- target
+      plan <- target$resources$plan %||% future::plan()
+      old_plan <- future::plan(plan)
+      on.exit(future::plan(old_plan))
       future <- future::future(
         expr = target_run_worker(.targets_target_5048826d),
         packages = "targets",